+++ project/src/main/java/com/Group4/MiniProject/Theme/entity/Theme.java
@@ -1,0 +1,20 @@
+package com.Group4.MiniProject.Theme.entity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Table(name = "theme")
+public class Theme {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    // 🛠TODO: change to STRING or ENUM TYPE
+    private Long themeId;
+    private String name;
+    private String imgUrl;
+}
