+++ project/src/main/java/com/Group4/MiniProject/Theme/repository/ThemeRepository.java
@@ -1,0 +1,7 @@
+package com.Group4.MiniProject.Theme.repository;
+
+import com.Group4.MiniProject.Theme.entity.Theme;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ThemeRepository extends JpaRepository<Theme, Long> {
+}
