+++ project/src/main/java/com/Group4/MiniProject/Theme/service/ThemeService.java
@@ -1,0 +1,23 @@
+package com.Group4.MiniProject.Theme.service;
+
+import com.Group4.MiniProject.Theme.entity.Theme;
+import com.Group4.MiniProject.Theme.dto.ThemeResponseDto;
+import com.Group4.MiniProject.Theme.repository.ThemeRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+@RequiredArgsConstructor
+public class ThemeService {
+
+    private final ThemeRepository themeRepository;
+
+    public List<ThemeResponseDto> getAllThemes() {
+        List<Theme> themes = themeRepository.findAll();
+
+        return themes.stream().map(ThemeResponseDto::new).collect(Collectors.toList());
+    }
+}
