+++ project/src/main/java/com/Group4/MiniProject/User/dto/UserRequestDto.java
@@ -1,0 +1,13 @@
+package com.Group4.MiniProject.User.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.AllArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserRequestDto {
+    private String nickname;
+    private String password;
+}
