+++ project/src/main/java/com/Group4/MiniProject/User/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.Group4.MiniProject.User.repository;
+
+import com.Group4.MiniProject.User.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByNickname(String nickname);
+
+    boolean existsByNickname(String nickname);
+}
