+++ project/src/main/java/com/Group4/MiniProject/common/dto/ErrorResponseDto.java
@@ -1,0 +1,10 @@
+package com.Group4.MiniProject.common.dto;
+
+import lombok.Getter;
+import lombok.AllArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+public class ErrorResponseDto {
+    private String error;
+}
