--- project/src/main/java/com/Group4/MiniProject/dto/ErrorResponseDto.java
@@ -1,10 +1,0 @@
-package com.Group4.MiniProject.dto;
-
-import lombok.Getter;
-import lombok.AllArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-public class ErrorResponseDto {
-    private String error;
-}
