--- project/src/main/java/com/Group4/MiniProject/dto/LoginResponseDto.java
@@ -1,15 +1,0 @@
-package com.Group4.MiniProject.dto;
-
-import lombok.Getter;
-import lombok.AllArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-public class LoginResponseDto {
-    private String message;
-    private String nickname;
-
-    public static LoginResponseDto success(String nickname) {
-        return new LoginResponseDto("로그인 성공", nickname);
-    }
-}
