--- project/src/main/java/com/Group4/MiniProject/dto/UserRequestDto.java
@@ -1,13 +1,0 @@
-package com.Group4.MiniProject.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserRequestDto {
-    private String nickname;
-    private String password;
-}
