--- project/src/main/java/com/Group4/MiniProject/repository/ThemeRepository.java
@@ -1,7 +1,0 @@
-package com.Group4.MiniProject.repository;
-
-import com.Group4.MiniProject.entity.Theme;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ThemeRepository extends JpaRepository<Theme, Long> {
-}
