--- project/src/main/java/com/Group4/MiniProject/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.Group4.MiniProject.repository;
-
-import com.Group4.MiniProject.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByNickname(String nickname);
-
-    boolean existsByNickname(String nickname);
-}
